po korektach

+++ ShootingRange/IShooter.cs
@@ -11,7 +11,7 @@ namespace ShootingRange
         void AddResult(string result);
         void AddResult(double result);
         void AddResult(char result);
-        List<int> GetResults();
+        List<float> GetResults();
         Statistics GetStatistics();
     }
 }
