+++ ShootingRange/ShooterInMemory.cs
@@ -2,14 +2,14 @@
 {
     public class ShooterInMemory : ShooterBase
     {
-        private List<int> results = new List<int>();
+        private List<float> results = new List<float>();
         public override event ResultAddedDelegate? ResultAdded;
 
         public ShooterInMemory() : base() {}
         
         public ShooterInMemory(string firstName, string lastName) : base(firstName, lastName) { }
         
-        public override void AddResult(int result)
+        public override void AddResult(float result)
         {
             if ((result >= Min) && (result <= Max))
             {
@@ -21,7 +21,7 @@
                 throw new Exception("Parameter out of range.");
         }
 
-        public override List<int> GetResults()
+        public override List<float> GetResults()
         {
             return results;
         }
